Maybe fix the exposemultiplexedframebytes branch

+++ mpifxcorr/src/vdiffile.h
@@ -91,7 +91,7 @@ protected:
   char formatname[64];
 
   unsigned char *readbuffer;
-  int readbuffersize;
+  long long readbuffersize;
   int readbufferslots;
   unsigned int readbufferslotsize;
   int readbufferleftover;
